+++ src/steps/project/index.test.ts
@@ -9,10 +9,10 @@ afterEach(async () => {
   await recording.stop();
 });
 
-test('fetch-projects', async () => {
+test(Steps.PROJECTS, async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'fetch-projects',
+    name: Steps.PROJECTS,
   });
 
   const stepConfig = buildStepTestConfigForStep(Steps.PROJECTS);
@@ -20,10 +20,21 @@ test('fetch-projects', async () => {
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });
 
-test('build-project-and-pipelines-relationships', async () => {
+test(Steps.PROJECTS_ENV_VARIABLES, async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'build-project-and-pipelines-relationships',
+    name: Steps.PROJECTS_ENV_VARIABLES,
+  });
+
+  const stepConfig = buildStepTestConfigForStep(Steps.PROJECTS_ENV_VARIABLES);
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
+});
+
+test(Steps.BUILD_USER_GROUPS_PROJECTS_RELATIONSHIPS, async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: Steps.BUILD_USER_GROUPS_PROJECTS_RELATIONSHIPS,
   });
 
   const stepConfig = buildStepTestConfigForStep(